requirements.txt
fastapi[all]==0.99.1
selenium==4.10.0
requests==2.31.0
SQLAlchemy==2.0.18
jdatetime==4.1.1
python-dotenv~=1.0.0